tests/wpt/tests/css/css-tables/parsing/caption-side-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Table Module Test: parsing caption-side with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-tables/#propdef-caption-side">
<meta name="assert" content="caption-side supports the full grammar 'top | bottom'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("caption-side", "top");
test_valid_value("caption-side", "bottom");
</script>
</body>
</html>
